Move tag strip into shared lib for csw and dcat
