+++ ckanext/dia/harvester/dcat.py
@@ -3,12 +3,12 @@ from string import Template
 import json
 
 import requests
-import re
 
 from ckan import model
 import ckan.plugins as plugins
 from ckan.logic.action.get import license_list
 from ckanext.dcat.harvesters import DCATJSONHarvester
+from ckanext.dia.converters import strip_invalid_tags_content
 from clean_frequency import clean_frequency
 
 log = getLogger(__name__)
@@ -153,20 +153,6 @@ class DIADCATJSONHarvester(DCATJSONHarvester):
         return package_dict, dcat_dict
 
 
-    def _strip_invalid_tags_content(self, tags):
-        '''Takes a list of tag dicts, converts invalid characters to spaces
-        and then removes any duplicate spaces.'''
-
-        def convert_tag(tag):
-            # Replace bad characters with spaces
-            tag['name'] = re.sub('[^\w|^\-|^\.]', ' ', tag['name'])
-            # Remove redundant spaces
-            tag['name'] = re.sub(' {2,}', ' ', tag['name'])
-            return tag
-
-        return map(convert_tag, tags)
-
-
     def modify_package_dict(self, package_dict, dcat_dict, harvest_object):
         try:
             conf = json.loads(harvest_object.source.config)
@@ -176,7 +162,7 @@ class DIADCATJSONHarvester(DCATJSONHarvester):
             conf = {}
 
         tags = package_dict.get('tags', [])
-        tags = self._strip_invalid_tags_content(tags)
+        tags = strip_invalid_tags_content(tags)
         tags.extend(conf.get('default_tags', []))
         package_dict['tags'] = dict((tag['name'], tag) for tag in tags).values()
